+++ packages/config-registry-controller/src/index.ts
@@ -24,8 +24,11 @@ export {
 export type {
   FetchConfigOptions,
   FetchConfigResult,
+  FetchEventsConfigResult,
   RegistryNetworkConfig,
   RegistryConfigApiResponse,
+  RegistryEventsConfig,
+  RegistryEventsConfigApiResponse,
 } from './config-registry-api-service/types.js';
 export type {
   ConfigRegistryApiServiceOptions,
@@ -35,6 +38,7 @@ export type {
 } from './config-registry-api-service/config-registry-api-service.js';
 export type {
   ConfigRegistryApiServiceFetchConfigAction,
+  ConfigRegistryApiServiceFetchEventsConfigAction,
   ConfigRegistryApiServiceMethodActions,
 } from './config-registry-api-service/config-registry-api-service-method-action-types.js';
 export type { NetworkFilterOptions } from './config-registry-api-service/filters.js';